Update README.md with guides and guidelines
To ease onboarding, I think it is important to add some short guides for getting started, to those who are not familiar with Github. Further, I want to add guidelines on how we intent to use the different features provided by Github, in Project Origin. These guidelines are meant for everyone, but will be collected here, to make it easier to find. 

I am working on creating these, and hence I am creating a PR instead of just committing directly.