Update Readme
I did some changes to this Readme, as a lot of progress has happened since its last edit. The main changes applied are:

- Added some missing guides and links in the overview of ressources
- Divided the available documents into new categories
- I removed the bullet-point contribution guidelines, as these instead will be located in a separate document, keeping the Readme concise, and obtain an easier overview.
- Added section on work structures and Trusted Committer as per agreed in weekly sync session 09-02-2023
Note: Currently setting up a nicely rendered docs, but until it is up and running this readme is the point of reference for navigation in the Origin-Collaboration repository.